Fix "unbound variable" bash error when OVERRIDE is not defined
By default, bash renders undefined variables as empty strings,
but entrypoint.sh does a "set -u" - which causes bash to report the
"unbound variable" error instead. That's incompatible with the way
the script checks if OVERRIDE is defined, by checking if its length is 0.

We now check the variable by doing "${OVERRIDE-}", where the "-" forces
the variable to be rendered as the empty string if not defined.